agent/requirements.txt
bedrock-agentcore
strands-agents[otel]
strands-agents-tools
mcp
httpx
aws-opentelemetry-distro
streamlit
